requirements.txt
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dlib==19.24.2
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.0
fonttools==4.47.0
idna==3.6
imageio==2.33.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
MarkupSafe==2.1.3
matplotlib==3.8.2
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pillow==10.2.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
setuptools==69.0.3
six==1.16.0
tifffile==2023.12.9
urllib3==2.1.0
Werkzeug==3.0.1
fastapi==0.68.2
uvicorn==0.15.0
